trading_bot_backend/requirements.txt
Alembic
SQLAlchemy
cryptography
fastapi
httpx
psycopg2-binary
python-binance
python-dotenv
python-jose[cryptography]
uvicorn
